+++ src/main/php/lang/reflection/Method.class.php
@@ -85,7 +85,14 @@ class Method extends Routine {
     // Compile signature
     $sig= '';
     foreach ($this->reflect->getParameters() as $i => $parameter) {
-      if ($t= $parameter->getType()) {
+      $t= $parameter->getType();
+      if ($t instanceof \ReflectionUnionType) {
+        $name= '';
+        foreach ($t->getTypes() as $component) {
+          $name.= '|'.$component->getName();
+        }
+        $type= substr($name, 1);
+      } else if ($t) {
         $type= strtr(PHP_VERSION_ID >= 70100 ? $t->getName() : $t->__toString(), '\\', '.');
         $parameter->isVariadic() && $type.= '...';
       } else if (isset($tags['param'][$i])) {
@@ -98,7 +105,14 @@ class Method extends Routine {
     }
 
     // Put together return type
-    if ($t= $this->reflect->getReturnType()) {
+    $t= $this->reflect->getReturnType();
+    if ($t instanceof \ReflectionUnionType) {
+      $name= '';
+      foreach ($t->getTypes() as $component) {
+        $name.= '|'.$component->getName();
+      }
+      $returns= substr($name, 1);
+    } else if ($t) {
       $returns= strtr(PHP_VERSION_ID >= 70100 ? $t->getName() : $t->__toString(), '\\', '.');
     } else if (isset($tags['return'][0])) {
       preg_match('/([^ ]+)( .+)?/i', $tags['return'][0], $matches);